apps/podcasts/pyproject.toml
[project]
name = "podcasts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Willem Meints", email = "1550763+wmeints@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dapr-ext-fastapi>=1.15.0",
    "dapr-ext-grpc>=1.15.0",
    "dapr>=1.15.0",
    "fastapi>=0.115.12",
    "opentelemetry-api>=1.33.1",
    "opentelemetry-exporter-zipkin-json>=1.33.1",
    "opentelemetry-instrumentation-fastapi>=0.54b1",
    "opentelemetry-sdk>=1.33.1",
    "pydantic>=2.10.5",
    "typer>=0.15.4",
    "uvicorn>=0.34.2",
]

[project.scripts]
agenttalks-podcasts = "agenttalks.podcasts.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agenttalks"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.10",
    "callee>=0.3.1",
]
